Rename menu with the name of the component
